requirements.txt
#pip3 install -r requirements.txt

python-dotenv #to manage environment variables
spotipy #python library for spotify api
pandas #data processing
requests #web requests
sendgrid==6.6.0 #email sender